[bug]Beim Loeschen von Schiffen wurde immer versucht die Skriptdaten zu entfernen, selbst wenn das Schiff gar keine hatte. -> NPE

+++ src/net/driftingsouls/ds2/server/ships/Ship.java
@@ -3562,7 +3562,10 @@ public class Ship implements Locatable,Transfering {
 			.setInteger(0, this.id)
 			.executeUpdate();
 		
-		db.delete(scriptData);
+		if(scriptData != null)
+		{
+			db.delete(scriptData);
+		}
 
 		db.delete(this);
